backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
numpy
torch
transformers
scikit-learn
together